Move v2 test fixtures to testing_helpers.go
