--- components/eventing-controller/api/v1alpha2/testing_helpers.go
@@ -1,20 +1,0 @@
-package v1alpha2
-
-var (
-	ProtocolSettingsContentModeBinary = "BINARY"
-	DefaultName                       = "test"
-	DefaultNamespace                  = "test-namespace"
-	DefaultSink                       = "https://svc2.test.local"
-	DefaultID                         = "id"
-	DefaultMaxInFlight                = 10
-	DefaultStatusReady                = true
-	DefaultConditions                 = []Condition{
-		{
-			Type:   ConditionSubscriptionActive,
-			Status: "true",
-		},
-		{
-			Type:   ConditionSubscribed,
-			Status: "false",
-		}}
-)
